Move loadProperties to its own module
