+++ validateProperties.js
@@ -1,29 +1,9 @@
-const { cloudCIOnlyMavenProperties, encoding } = require("./constants");
+const { cloudCIOnlyMavenProperties } = require("./constants");
 const fs = require("fs");
 const path = require("path");
-const os = require("os");
 const assert = require("./assert");
 const sensitive = require("./sensitive");
-
-// Loads a Java .properties file into a Map.
-const loadProperties = fileName =>
-  fs
-    .readFileSync(fileName, encoding)
-    .split(os.EOL)
-    .reduce((acc, cur) => {
-      // Any space between the key and value is removed.
-      // Any trailing space is left as is.
-      cur = cur.trimLeft();
-
-      if (!cur.startsWith("#") && cur.includes("=")) {
-        // Only match the first equals sign.
-        // https://stackoverflow.com/questions/4607745/split-string-only-on-first-instance-of-specified-character
-        let items = cur.split(/=(.*)/);
-        acc.set(items[0].trim(), items[1].trimLeft());
-      }
-
-      return acc;
-    }, new Map());
+const { loadProperties } = require("./loadProperties");
 
 const validateProperties = (folderInfo, pomInfo) => {
   let localProperties = loadProperties(folderInfo.localPropertiesFile);
